Refactor `SystemTransitionType`
