+++ IBAnimatable/AnimatedTransitioning.swift
@@ -25,7 +25,7 @@ public protocol AnimatedTransitioning: ViewControllerAnimatedTransitioning {
 }
 
 public extension AnimatedTransitioning {
-  public func animateWithCATransition(transitionContext: UIViewControllerContextTransitioning, type: SystemTransitionType, subtype: String?) {
+  public func animateWithCATransition(transitionContext: UIViewControllerContextTransitioning, type: TransitionAnimationType.SystemTransitionType, subtype: String?) {
     let (_, tempToView, tempContainerView) = retrieveViews(transitionContext: transitionContext)
     guard let toView = tempToView, let containerView = tempContainerView else {
       transitionContext.completeTransition(true)
