+++ IBAnimatable/SystemPageAnimator.swift
@@ -40,9 +40,9 @@ extension SystemPageAnimator: UIViewControllerAnimatedTransitioning {
   public func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
     switch self.type {
     case .curl:
-      animateWithCATransition(transitionContext: transitionContext, type: SystemTransitionType.pageCurl, subtype: nil)
+      animateWithCATransition(transitionContext: transitionContext, type: TransitionAnimationType.SystemTransitionType.pageCurl, subtype: nil)
     case .unCurl:
-      animateWithCATransition(transitionContext: transitionContext, type: SystemTransitionType.pageUnCurl, subtype: nil)
+      animateWithCATransition(transitionContext: transitionContext, type: TransitionAnimationType.SystemTransitionType.pageUnCurl, subtype: nil)
     }
   }
 }
