+++ IBAnimatable/SystemSuckEffectAnimator.swift
@@ -26,6 +26,6 @@ extension SystemSuckEffectAnimator: UIViewControllerAnimatedTransitioning {
   }
   
   public func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
-    animateWithCATransition(transitionContext: transitionContext, type: SystemTransitionType.suckEffect, subtype: nil)
+    animateWithCATransition(transitionContext: transitionContext, type: TransitionAnimationType.SystemTransitionType.suckEffect, subtype: nil)
   }
 }
